// air-ops-v2/src/features/operator/types/type.ts
import { FileObject } from "@/shared/types/common";

export type OperatorFormValues = {
  name: string;
  email: string;
  phone: string;
  address: string;
  city: string;
  state: string;
  country: string;
  pinCode: string;
  companyName: string;
  companyAddress: string;
  companyEmail: string;
  companyPhone: string;
  branch: string;
  companyLogo: FileObject | null;
  supportEmail: string;
  ticketFooterNote: string;
  websiteUrl: string;
  themeColor?: string;
  subscriptionPlan?: string;
  billingCycle?: string;
};